+++ packages/gatsby-plugin-manifest/src/__tests__/gatsby-node.js
@@ -10,6 +10,10 @@ const path = require(`path`)
 const { onPostBootstrap } = require(`../gatsby-node`)
 
 describe(`Test plugin manifest options`, () => {
+  beforeEach(() => {
+    fs.writeFileSync.mockReset()
+  })
+
   it(`correctly works with default parameters`, async () => {
     await onPostBootstrap([], {
       name: `GatsbyJS`,
@@ -23,7 +27,8 @@ describe(`Test plugin manifest options`, () => {
     expect(filePath).toEqual(path.join(`public`, `manifest.webmanifest`))
     expect(contents).toMatchSnapshot()
   })
-  it(`fails on non existing icon`, (done) => {
+
+  it(`fails on non existing icon`, done => {
     fs.statSync.mockReturnValueOnce({ isFile: () => false })
     onPostBootstrap([], {
       name: `GatsbyJS`,
@@ -33,15 +38,46 @@ describe(`Test plugin manifest options`, () => {
       theme_color: `#a2466c`,
       display: `minimal-ui`,
       icon: `non/existing/path`,
-      icons: [{
-        src: `icons/icon-48x48.png`,
-        sizes: `48x48`,
-        type: `image/png`,
-      }],
-    })
-    .catch((err) => {
+      icons: [
+        {
+          src: `icons/icon-48x48.png`,
+          sizes: `48x48`,
+          type: `image/png`,
+        },
+      ],
+    }).catch(err => {
       expect(err).toMatchSnapshot()
       done()
     })
+  })
+
+  it(`doesn't write extra properties to manifest`, async () => {
+    const manifestOptions = {
+      name: `GatsbyJS`,
+      short_name: `GatsbyJS`,
+      start_url: `/`,
+      background_color: `#f7f0eb`,
+      theme_color: `#a2466c`,
+      display: `minimal-ui`,
+      icons: [
+        {
+          src: `icons/icon-48x48.png`,
+          sizes: `48x48`,
+          type: `image/png`,
+        },
+      ],
+    }
+    const pluginSpecificOptions = {
+      icon: undefined,
+      legacy: true,
+      plugins: [],
+    }
+    await onPostBootstrap([], {
+      ...manifestOptions,
+      ...pluginSpecificOptions,
+    })
+
+    const content = JSON.parse(fs.writeFileSync.mock.calls[0][1])
+    expect(content).toEqual(manifestOptions)
   })
 })