+++ packages/gatsby-plugin-google-gtag/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.5"></a>
+
+## [1.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.4...gatsby-plugin-google-gtag@1.0.5) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
+<a name="1.0.4"></a>
+
+## [1.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.3...gatsby-plugin-google-gtag@1.0.4) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
+<a name="1.0.3"></a>
+
+## [1.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.2...gatsby-plugin-google-gtag@1.0.3) (2018-11-01)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
+<a name="1.0.2"></a>
+
+## [1.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.1...gatsby-plugin-google-gtag@1.0.2) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
 <a name="1.0.1"></a>
 
 ## [1.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.0...gatsby-plugin-google-gtag@1.0.1) (2018-10-19)